crew_AI_tinyllama/requirements.txt
crewai
python-dotenv
crewai-tools
requests 